added user

+++ src/Backend/Jobs/Application/ApiCallJobCommandQuery/Commands/Delete/DeleteApiCallJobCommand.cs
@@ -8,10 +8,12 @@ public record DeleteApiCallJobCommand(int Id) : IRequest;
 public class DeleteShipCommandHandler : IRequestHandler<DeleteApiCallJobCommand>
 {
     private readonly IApplicationDbContext _context;
+    private readonly ICurrentUserService _currentUserService;
 
-    public DeleteShipCommandHandler(IApplicationDbContext context)
+    public DeleteShipCommandHandler(IApplicationDbContext context, ICurrentUserService currentUserService)
     {
         _context = context;
+        _currentUserService = currentUserService;
     }
 
     public async Task Handle(DeleteApiCallJobCommand request, CancellationToken cancellationToken)
@@ -20,11 +22,11 @@ public class DeleteShipCommandHandler : IRequestHandler<DeleteApiCallJobCommand>
         var entity = await _context.ApiCallJobs
           .FindAsync(new object[] { request.Id }, cancellationToken)
           ?? throw new NotFoundException(nameof(ApiCallJob), request.Id);
-
-        entity.AddDomainEvent(new ApiCallJobDeletedEvent(entity));
-
-        _context.ApiCallJobs.Remove(entity);
-
-        await _context.SaveChangesAsync(cancellationToken);
+        if (entity.UserId == Guid.Parse(_currentUserService.UserId!))
+        {
+            entity.AddDomainEvent(new ApiCallJobDeletedEvent(entity));
+            _context.ApiCallJobs.Remove(entity);
+            await _context.SaveChangesAsync(cancellationToken);
+        }
     }
 }
